+++ src/components/Card/StyledCard.tsx
@@ -1,0 +1,6 @@
+import { styled } from "styled-components";
+
+const StyledCard = styled.div`
+  width: 30%;
+`;
+export default StyledCard;