+++ src/components/Stack/Stack.tsx
@@ -1,0 +1,12 @@
+import { ReactNode } from "react";
+import StackStyled from "./StackStyled";
+
+interface StackProps {
+  children: ReactNode;
+}
+
+const Stack = ({ children }: StackProps) => {
+  return <StackStyled>{children}</StackStyled>;
+};
+
+export default Stack;