+++ src/components/Stack/StackStyled.tsx
@@ -1,0 +1,10 @@
+import { styled } from "styled-components";
+
+const StackStyled = styled.div`
+  margin: 2rem 0;
+  display: flex;
+  justify-content: space-around;
+  flex-wrap: wrap;
+`;
+
+export default StackStyled;