+++ src/lib/endpoints/endpoints.ts
@@ -1,0 +1,9 @@
+
+const BASE_URL = "https://covid-fe-2023.vercel.app/api";
+
+const ENDPOINTS = {
+  GLOBAL: `${BASE_URL}/global.json`,
+  INDONESIA: `${BASE_URL}/indonesia.json`
+};
+
+export default ENDPOINTS;