+++ src/lib/store/store.ts
@@ -1,0 +1,11 @@
+import { configureStore } from "@reduxjs/toolkit";
+import provincesReducer from "../slices/provincesSlice";
+
+const store = configureStore({
+  reducer: {
+    provinces: provincesReducer,
+  },
+});
+
+export type RootState = ReturnType<typeof store.getState>;
+export default store;