+++ CHANGELOG.md
@@ -2,6 +2,69 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.0.3](https://github.com/amar-ti2021/uas-frontend/compare/v0.0.2...v0.0.3) (2023-07-07)
+
+
+### Features
+
+* Add axios package for HTTP requests ([21312a5](https://github.com/amar-ti2021/uas-frontend/commit/21312a51228705f750d12f8e07da4fd065bc2898))
+* Add Footer component ([6e98ca6](https://github.com/amar-ti2021/uas-frontend/commit/6e98ca6d20e4be2ae62a62747796717fc9900780))
+* Add Hero, Outlet, and Footer to Layout component ([40bd68c](https://github.com/amar-ti2021/uas-frontend/commit/40bd68c7ba0a9bc653754587d6e82d212a5f48a0))
+* Add new routes for global, Indonesia, provinsi, and about pages ([5cb41eb](https://github.com/amar-ti2021/uas-frontend/commit/5cb41eb6e2455d3b846b5961b7dbe250db4cefa9))
+* Add react-redux Provider to App.tsx ([11ebd87](https://github.com/amar-ti2021/uas-frontend/commit/11ebd873929080d8bd0ab460b200a9c815c2d68c))
+* Add Redux Toolkit Store ([9f08aea](https://github.com/amar-ti2021/uas-frontend/commit/9f08aea0779aa2a71a29a4c7525fa23e17f7dc09))
+* Add SVG file for Hero component ([e65073d](https://github.com/amar-ti2021/uas-frontend/commit/e65073dbfcea20a75ef2f6e799473189e46a31ba))
+* Create About Page ([6b78aed](https://github.com/amar-ti2021/uas-frontend/commit/6b78aed14b06b23a0d1721a666c5cff193e4f72b))
+* Create Card component for displaying content ([748b0b6](https://github.com/amar-ti2021/uas-frontend/commit/748b0b653716965db417d2c89611abcec5a49067))
+* Create endpoints constant for managing API endpoints ([75bc625](https://github.com/amar-ti2021/uas-frontend/commit/75bc6251d19c0e3be9948213f35e7ee7bbb8ebc9))
+* Create Global Page ([f0521a7](https://github.com/amar-ti2021/uas-frontend/commit/f0521a71cfb0a2fc4ac865e5e412b45949c78bc4))
+* Create GlobalSection component ([1d67549](https://github.com/amar-ti2021/uas-frontend/commit/1d67549d77c98db20078f8ae67ff02e0086f57ce))
+* Create Hero component ([f43d38b](https://github.com/amar-ti2021/uas-frontend/commit/f43d38bfa8964272e72e20e94dbc00f3b4d8234e))
+* Create Indonesia Page ([7471cfe](https://github.com/amar-ti2021/uas-frontend/commit/7471cfec26fc830337304291ef97466dce84fb6d))
+* Create IndonesiaSection component ([ca03a5c](https://github.com/amar-ti2021/uas-frontend/commit/ca03a5cdfdcbaed46efcd09d6004ae54e0910c33))
+* Create provinces constant for provincesSlice requiring local data ([aa5bb65](https://github.com/amar-ti2021/uas-frontend/commit/aa5bb65427663804b16f03d20f537ef06fcd149a))
+* Create ProvinceSection component ([24b5921](https://github.com/amar-ti2021/uas-frontend/commit/24b59213b452fd7523afe9567ccab3ca28aeb9ba))
+* Create provincesSlice ([6a3a04c](https://github.com/amar-ti2021/uas-frontend/commit/6a3a04c30af21390486e9a9724d4b39914dbde33))
+* Create RegionSection component ([c97aa7e](https://github.com/amar-ti2021/uas-frontend/commit/c97aa7e576d99ab14a7ead065cf18b7c15fd7123))
+* Create Section component as a container for sections ([98631d9](https://github.com/amar-ti2021/uas-frontend/commit/98631d97198ccba9144162a3e7cbcdee0484eaa2))
+* Create Stack component as a flexible container for components ([55b6ff0](https://github.com/amar-ti2021/uas-frontend/commit/55b6ff0174d16549a73129cf0327266fc5e4d572))
+* Create StyledCard component to style the Card component ([36206e8](https://github.com/amar-ti2021/uas-frontend/commit/36206e8b63e7dd8ac381c5b01cdfbbe20ada773d))
+* Create Subtitle Component ([0402e5a](https://github.com/amar-ti2021/uas-frontend/commit/0402e5a4c20aa171f4eaebd5f968d8187f67003e))
+* Create Table component as a custom table for displaying data ([fa14e1d](https://github.com/amar-ti2021/uas-frontend/commit/fa14e1d39ded7bd3c1f8e9c00e47a2164aed8724))
+* Create Title Component ([b907055](https://github.com/amar-ti2021/uas-frontend/commit/b907055bf7ae9a2c885e56525297452902dc9360))
+* Create UpdateForm component for modifying province data ([2ef4ec6](https://github.com/amar-ti2021/uas-frontend/commit/2ef4ec6ac028ec4f06581ad073f776eb3876c9e5))
+* Create useTransformProvinceData hook for transforming local data into a common province data format ([5f4be4b](https://github.com/amar-ti2021/uas-frontend/commit/5f4be4bf1915416bc6fc5e4faefb3651fa5d6927))
+* Integrate react-redux and redux-toolkit ([a526c54](https://github.com/amar-ti2021/uas-frontend/commit/a526c54eff5d68eaf35bcb88291af269e6ef734e))
+* Remove 'allowSyntheticDefaultImports' import from tsconfig.json ([8454b51](https://github.com/amar-ti2021/uas-frontend/commit/8454b51486f609e5f5fd59a4654455a505964431))
+* Set max-width of the body to improve layout ([099f490](https://github.com/amar-ti2021/uas-frontend/commit/099f490247eb6c75ff39e7232ce4706124d1eb74))
+* Update background color to a lighter shade of dark blue ([940286c](https://github.com/amar-ti2021/uas-frontend/commit/940286c9a5eca94cc70c2a2f3e24d932ca5452fb))
+
+
+### Bug Fixes
+
+* Remove 'import App.css' causing redundant import error ([2e48767](https://github.com/amar-ti2021/uas-frontend/commit/2e4876734604f4c0ad7d3e01fd29232f22fa854f))
+
+### [0.0.2](https://github.com/amar-ti2021/uas-frontend/compare/v0.0.1...v0.0.2) (2023-07-05)
+
+
+### Features
+
+* Add Navbar component to Layout ([d83325a](https://github.com/amar-ti2021/uas-frontend/commit/d83325a2cbbb764b0d57f5644e3e71725db18c7c))
+* add react-router-dom package ([c851365](https://github.com/amar-ti2021/uas-frontend/commit/c85136549c8abebe03b565416c9534f082d60e00))
+* add RouterProvider ([f98a421](https://github.com/amar-ti2021/uas-frontend/commit/f98a4216498cb1ca1cf501ac1e9810d21246668c))
+* add styled-components package ([0f61961](https://github.com/amar-ti2021/uas-frontend/commit/0f61961b69e9d8517c64f84d6e57dd5148894723))
+* create Browser Router "router" ([6ea87cf](https://github.com/amar-ti2021/uas-frontend/commit/6ea87cf78fa67f49a1caf18cf48278d8dd3592ae))
+* create Layout component ([82b3a46](https://github.com/amar-ti2021/uas-frontend/commit/82b3a4671b264990d98088291cceae36af356e12))
+* Create Navbar component ([bb2754a](https://github.com/amar-ti2021/uas-frontend/commit/bb2754a90507ee2a8c1e6f4d2f8747b7b783d4d5))
+* Create StyledNavbar component using styled-components ([7bf9747](https://github.com/amar-ti2021/uas-frontend/commit/7bf974778b27dc57b5ff1bed187fedc0e5316eb4))
+
+
+### Bug Fixes
+
+* Fix import error caused by .tsx file extension in App import ([774785c](https://github.com/amar-ti2021/uas-frontend/commit/774785c17c74e10d85955394e58fa1e3b71398bc))
+* tsconfig moduleResolution errors ([6f48635](https://github.com/amar-ti2021/uas-frontend/commit/6f486358b3cbb9da7a15cdfc2e328d691b9ea508))
+* tsconfig synthetic default imports ([3141d78](https://github.com/amar-ti2021/uas-frontend/commit/3141d78b304e27028adb96d4701f6b09c8e29ae1))
+
 ### 0.0.1 (2023-07-03)